Fraction->double in SfxViewShell::SetZoomFactor
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202135
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202798
Tested-by: Tomaž Vajngerl <quikee@gmail.com>

+++ sc/source/ui/view/gridwin2.cxx
@@ -262,7 +262,7 @@ void ScGridWindow::DoPushPivotToggle( SCCOL nCol, SCROW nRow, const MouseEvent& 
     sal_uInt16 nIndent = rIndentItem.GetValue();
 
     // Check if the mouse cursor is clicking on the toggle +/- box.
-    ScDPFieldButton aBtn(GetOutDev(), GetSettings().GetStyleSettings(), GetMapMode().GetScaleY(), mrViewData.GetDocument());
+    ScDPFieldButton aBtn(GetOutDev(), GetSettings().GetStyleSettings(), double(GetMapMode().GetScaleY()), mrViewData.GetDocument());
     aBtn.setBoundingBox(aScrPos, aScrSize, bLayoutRTL);
     aBtn.setDrawToggleButton(true, true, nIndent);
     Point aPopupPos;
@@ -407,7 +407,7 @@ bool ScGridWindow::DPTestFieldPopupArrow(
     Size aScrSize(nSizeX-1, nSizeY-1);
 
     // Check if the mouse cursor is clicking on the popup arrow box.
-    ScDPFieldButton aBtn(GetOutDev(), GetSettings().GetStyleSettings(), GetMapMode().GetScaleY(), mrViewData.GetDocument());
+    ScDPFieldButton aBtn(GetOutDev(), GetSettings().GetStyleSettings(), double(GetMapMode().GetScaleY()), mrViewData.GetDocument());
     aBtn.setBoundingBox(aScrPos, aScrSize, bLayoutRTL);
     aBtn.setPopupLeft(false);   // DataPilot popup is always right-aligned for now
     Point aPopupPos;
@@ -437,7 +437,7 @@ bool ScGridWindow::DPTestMultiFieldPopupArrow(
     Size aScrSize(nSizeX - 1, nSizeY - 1);
 
     // Check if the mouse cursor is clicking on the popup arrow box.
-    ScDPFieldButton aBtn(GetOutDev(), GetSettings().GetStyleSettings(), GetMapMode().GetScaleY(), mrViewData.GetDocument());
+    ScDPFieldButton aBtn(GetOutDev(), GetSettings().GetStyleSettings(), double(GetMapMode().GetScaleY()), mrViewData.GetDocument());
     aBtn.setBoundingBox(aScrPos, aScrSize, bLayoutRTL);
     aBtn.setPopupLeft(false);   // DataPilot popup is always right-aligned for now
     aBtn.setDrawPopupButtonMulti(true);
